Turn input string into binary matrix
Each row represents a `char *` representation of the base-2 ASCII
equivalent of a given character in the input string.

Going forward, we need to make this even (doing left-padding), so that
all rows will have equal width. We ultimately need this to be able to decode
properly and also for the next step when we further encode to triplets.